backend/requirements.txt
flask
flask-cors
torch
Pillow
open-clip-torch